requirement.txt

attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==4.1.0
bs4==0.0.1

defusedxml==0.7.1
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
Flask==2.2.5
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1

importlib-metadata==6.7.0
importlib-resources==5.2.0

Jinja2==3.1.2

lxml==4.9.3
matplotlib==3.5.3
matplotlib-inline==0.1.6

nbclient==0.5.13
nbconvert==6.4.4

numpy==1.21.6
openpyxl==3.0.10
packaging==22.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3

pip==22.3.1

PyJWT==2.8.0
pymongo==4.5.0
pyparsing==3.1.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1

regex==2022.7.9
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
six==1.16.0

SQLAlchemy==2.0.21

urllib3==2.0.6
waitress==2.1.2
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
wheel==0.38.4
WTForms==3.0.1
xlrd==2.0.1
zipp==3.15.0
